Loading in data and model architecture rebuilding

+++ autoencoder.py
@@ -8,7 +8,7 @@ import keras
 from keras.layers import Dense, Flatten, Input
 from keras.datasets import mnist
 import numpy as np
-
+import os
 
 # In[ ]:
 
@@ -72,22 +72,38 @@ def addNoise(images, multiplier=2):
 
 model = build_model(middle_nodes=28)
 print('model built')
-(x_train, x_test) = getRegularData()
-print('did it just become 0?', len(x_train))
-print('regular data aquired')
 
-(noisy_x_train, noisy_x_test) = getNoisyData(x_train, x_test)
-print('noisy data aquired')
-print('noisy data len:', len(noisy_x_train))
-print('reg data len:', len(x_train))
-noisy_x_train = np.array(noisy_x_train)
-noisy_x_test = np.array(noisy_x_test)
-x_train = np.array(x_train)
-x_test = np.array(x_test)
-np.save('noisy_x_train.npy', noisy_x_train)
-np.save('noisy_x_test.npy', noisy_x_test)
-np.save('x_train.npy', x_train)
-np.save('x_test.npy', x_test)
+x_train = None
+x_test = None
+noisy_x_train = None
+noisy_x_test = None
+if os.path.exists('noisy_x_train.npy') and os.path.exists('noisy_x_test.npy') and os.path.exists('x_train.npy') and os.path.exists('x_test.npy'):
+    x_train = np.load('x_train.npy')
+    x_test = np.load('x_test.npy')
+    noisy_x_train = np.load('noisy_x_train.npy')
+    noisy_x_test = np.load('noisy_x_test.npy')
+else:
+    (x_train, x_test) = getRegularData()
+    print('did it just become 0?', len(x_train))
+    print('regular data aquired')
+
+    (noisy_x_train, noisy_x_test) = getNoisyData(x_train, x_test)
+    print('noisy data aquired')
+    print('noisy data len:', len(noisy_x_train))
+    print('reg data len:', len(x_train))
+    noisy_x_train = np.array(noisy_x_train)
+    noisy_x_test = np.array(noisy_x_test)
+    x_train = np.array(x_train)
+    x_test = np.array(x_test)
+    np.save('noisy_x_train.npy', noisy_x_train)
+    np.save('noisy_x_test.npy', noisy_x_test)
+    np.save('x_train.npy', x_train)
+    np.save('x_test.npy', x_test)
+
+x_train = np.reshape(x_train, (-1, 28, 28, 1))
+x_test = np.reshape(x_test, (-1, 28, 28, 1))
+noisy_x_train = np.reshape(noisy_x_train, (-1, 28, 28, 1))
+noisy_x_test = np.reshape(noisy_x_test, (-1, 28, 28, 1))
 model.fit(x=noisy_x_train, y=x_train, batch_size=500, epochs=10, verbose=2, validation_data=(noisy_x_test, x_test))
